feat: implement command alias backend
The alias will basically be a macro / shortcut, rather than a strict command <-> alias sharing of functionality. So given `!alias pa points add`, the following would be exactly the same:

```
!points add citycide 50
!pa citycide 50
```

Trailing strings will remain as-is, which allows passing through arguments as if the original command was used normally.

There's will be no difference in the event object, so modules won't be able to tell the difference between the alias being used vs the original command.